refactor(kanji): rename zenz.rs and kanji_zenz_smoke.rs to llama_cpp variants (pure git mv)
File rename via git mv so the rename detection threshold stays high. The
bodies of the renamed files are untouched in this commit; struct / impl
/ docstring rewrites follow in the next commit.

mod.rs is also updated to reference the new module name and feature gate
(zenz -> llama-cpp). Without this single-line pair update, cargo fmt
--all --check (invoked by the pre-commit hook) cannot parse the module
tree because 'mod zenz;' no longer resolves. The struct re-export still
names the old ZenzBackend symbol here because the rename happens in the
next commit.

Task: P1-2.5-5 (step A of 2-commit split per plan review M1)
Ref: #73
