+++ crates/kotoha-core/tests/kanji_llama_cpp_smoke.rs
